Fixed async batch llm operations, password hashing

+++ src/data/session_manager.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from datetime import datetime
 from local import constants
 
-from sqlalchemy import create_engine, Column, Integer, String
+from sqlalchemy import create_engine, Column, Integer, String, Boolean
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy_utils import EncryptedType
@@ -17,24 +17,14 @@ Base = declarative_base()
 class User(Base):
     __tablename__ = 'users'    
     id = Column(Integer, primary_key=True)
-    name = Column(String)
-    password = Column(EncryptedType(String, 'password', AesEngine, 'pkcs5'))
+    name = Column(String, unique=True)
+    password = Column(String)
+    online = Column(Boolean)
     openai_api_key = Column(EncryptedType(String, 'openai_api_key', AesEngine, 'pkcs5'))
     role = Column(String)
     created_at = Column(String)
     last_login = Column(String)
-    time_zone = Column(String)
-    preferred_output_dir = Column(String)
-    preferred_pronouns = Column(String)
-    avatar_path = Column(String)
-    dark_mode = Column(String)
-    preferred_llm = Column(String)
-    preferred_llm_temperature = Column(String)
-    chat_color = Column(String)
-    bio = Column(String)
-    email = Column(String)
-    phone = Column(String)
-    website = Column(String)
+
 
 class SessionManager:
     """Manages the creation, reading, and writing of files and databases."""
@@ -44,29 +34,17 @@ class SessionManager:
         # print("Pwd context: " + str(self.pwd_context))
         # Setup session variables
         self.start_time = datetime.now()
-        # base directory should be the project directory's local folder
         self.base_dir = os.path.join(os.path.dirname(os.path.dirname(os.path.dirname(__file__))), 'local')
-        # print("Base dir: " + self.base_dir)
         self.docs_dir = Path.home() / 'Documents' / app_name
-        # print("Docs dir: " + str(self.docs_dir))
         self.pref_dir = self.base_dir    
-        # print("Pref dir: " + str(self.pref_dir))   
         self.users_db_path = os.path.join(self.base_dir, 'users.db')
-        # print("Users db path: " + self.users_db_path)
         self.encryption_key = key
-        # print("Encryption key: " + str(self.encryption_key))
         self.current_user = None
-        # print("Current user: " + str(self.current_user))
-        self.current_user_name = constants.DEFAULT_USER_NAME # Just for testing   
-        # print("Current user name: " + str(self.current_user_name))     
+        self.current_user_name = constants.DEFAULT_USER_NAME # Just for testing 
         self.preferred_output_dir = None
-        # print("Preferred output dir: " + str(self.preferred_output_dir))
         self.engine = None
-        # print("Engine: " + str(self.engine))
         self.session = None
-        # print("Session: " + str(self.session))
         self.setup()
-        # print("Setup complete")
 
     ####################################################################################################
     # Database operations
@@ -80,13 +58,10 @@ class SessionManager:
         if not os.path.exists(self.base_dir):
             # Create the base directory
             self.ensure_dir_exists(self.base_dir)
-            # Create the users database
-
-        # print(str(self.base_dir))
+       
         # Check if the engine is already created
         if self.engine is None:
             # Create the engine
-            # print("Creating engine")
             self.engine = create_engine('sqlite:///' + self.users_db_path)        
         Base.metadata.create_all(self.engine)
         Session = sessionmaker(bind=self.engine)
@@ -94,43 +69,26 @@ class SessionManager:
 
         # if no users exist, create a default admin user
         if self.session.query(User).count() == 0:
-            # print("No users found.")
-            self.create_user("admin", "admin", "pasword", "None", os.path.join(self.base_dir, "private/admin"), "he/him")
-            
-        else:
-            # print("Users exist")
-            # Let's print out all the users in the database by name
-            for user in self.session.query(User).all():
-                # print(user.name)
-                pass
+            self.create_user("admin", "admin", "password", "None")
         
         
-            
-
-    def create_user(self, name, role, password, openai_api_key, preferred_output_dir, preferred_pronouns):
+    def create_user(self, name, role, password, openai_api_key):
         """Creates a new user and adds them to the users database."""
         created_at = datetime.now().strftime("%m/%d/%Y, %H:%M:%S")
         last_login = datetime.now().strftime("%m/%d/%Y, %H:%M:%S")
-        time_zone = datetime.now().strftime("%Z")
-        preferred_output_dir = os.path.join(self.base_dir, name)
         hashed_password = self.pwd_context.hash(password)
-        # When verifying the password during a login, you can use pwd_context.verify(input_password, stored_password)
-        # which will return True if the input_password matches the stored_password after hashing.
 
         user = User(name=name, 
                     password=hashed_password, 
                     openai_api_key=openai_api_key, 
+                    online=False,
                     role=role, 
                     created_at=created_at, 
-                    last_login=last_login, 
-                    time_zone=time_zone, 
-                    preferred_output_dir=preferred_output_dir, 
-                    preferred_pronouns=preferred_pronouns)
+                    last_login=last_login)
         self.session.add(user)
         self.session.commit()
 
         self.current_user = user
-        self.preferred_output_dir = user.preferred_output_dir
 
         # Create the user's output directories
         self.ensure_dir_exists(user.preferred_output_dir)
@@ -138,6 +96,7 @@ class SessionManager:
         self.ensure_dir_exists(str(os.path.join(self.preferred_output_dir, "logs")))
         self.ensure_dir_exists(str(os.path.join(self.preferred_output_dir, "models")))
         self.ensure_dir_exists(str(os.path.join(self.preferred_output_dir, "transcripts")))
+
     
     def update_user(self, name, field, new_value):
         """Updates the specified field for a user."""
@@ -160,6 +119,11 @@ class SessionManager:
         """Logs out the current user."""
         self.current_user_name = None
 
+    def is_admin(self, name):
+        """Checks if a user is an admin."""
+        user = self.session.query(User).filter_by(name=name).first()
+        return user.role == 'admin'
+
     def get_all_users(self):
         """Returns all users in the database."""
         return self.session.query(User).all()
@@ -171,10 +135,21 @@ class SessionManager:
     
     def delete_user(self, name):
         """Deletes a user from the database."""
+        if not self.is_admin(self.current_user_name):
+            raise Exception('Only admins can delete users.')
         user = self.session.query(User).filter_by(name=name).first()
         if user:
             self.session.delete(user)
             self.session.commit()
+
+    def reset_password(self, name, new_password):
+        """Resets a user's password."""
+        user = self.session.query(User).filter_by(name=name).first()
+        if user:
+            hashed_password = self.pwd_context.hash(new_password)
+            user.password = hashed_password
+            self.session.commit()
+
